lob-quant-hft/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.11.0
statsmodels>=0.14.0
numba>=0.57.0
jupyter>=1.0.0
ipykernel>=6.25.0
plotly>=5.15.0
h5py>=3.9.0
pyarrow>=12.0.0
fastparquet>=2023.7.0
ta>=0.10.2
